Added stage on CustomPosts
